CollegeSearch/src/web/HelloController.java
package web;

import java.util.Date;

import org.apache.commons.logging.Log;
import org.apache.commons.logging.LogFactory;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.servlet.ModelAndView;

@Controller
public class HelloController {

	protected final Log logger = LogFactory.getLog(getClass());

	@RequestMapping({ "/", "/hello" })
	public ModelAndView sayHello() {
		return new ModelAndView("hello", "message", "Hello World");
	}

	@RequestMapping(value = "/welcome", method = RequestMethod.GET)
	public String welcome(Model model) {
		Date today = new Date();
		model.addAttribute("today", today);
		return "welcome";

	}
}